+++ src/mixins/itext_click_behavior.mixin.ts
@@ -111,7 +111,7 @@ export abstract class ITextClickBehaviorMixin<
       return;
     }
 
-    this.__isMousedown = true;
+    this.canvas.textEditingManager.register(this);
 
     if (this.selected) {
       this.inCompositionMode = false;
@@ -172,7 +172,17 @@ export abstract class ITextClickBehaviorMixin<
    * @private
    */
   mouseUpHandler(options: TransformEvent) {
-    this.__isMousedown = false;
+    if (this.canvas) {
+      this.canvas.textEditingManager.unregister(this);
+
+      const activeObject = this.canvas._activeObject;
+      if (activeObject && activeObject !== this) {
+        // avoid running this logic when there is an active object
+        // this because is possible with shift click and fast clicks,
+        // to rapidly deselect and reselect this object and trigger an enterEdit
+        return;
+      }
+    }
     if (
       !this.editable ||
       (this.group && !this.group.interactive) ||
@@ -180,16 +190,6 @@ export abstract class ITextClickBehaviorMixin<
       (options.e.button && options.e.button !== 1)
     ) {
       return;
-    }
-
-    if (this.canvas) {
-      const currentActive = this.canvas._activeObject;
-      if (currentActive && currentActive !== this) {
-        // avoid running this logic when there is an active object
-        // this because is possible with shift click and fast clicks,
-        // to rapidly deselect and reselect this object and trigger an enterEdit
-        return;
-      }
     }
 
     if (this.__lastSelected && !this.__corner) {